requirements.txt
absl-py==1.3.0
addict==2.4.0
appdirs==1.4.4
asttokens==2.0.5
backcall==0.2.0
black==22.10.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cython==0.29.32
debugpy==1.5.1
decorator==5.1.1
decord==0.6.0
distlib==0.3.6
docker-pycreds==0.4.0
easydict==1.10
entrypoints==0.4
executing==0.8.3
filelock==3.9.0
filterpy==1.4.5
flask-cors==3.0.10
flask==2.2.2
flatbuffers==22.12.6
fonttools==4.38.0
fvcore==0.1.5.post20221213
gitdb==4.0.10
gitpython==3.1.31
google-auth-oauthlib==0.4.6
google-auth==2.14.1
grpcio==1.50.0
huggingface-hub==0.12.1
humanfriendly==10.0
identify==2.5.13
idna==3.4
imageio-ffmpeg==0.4.7
imageio==2.22.4
importlib-metadata==5.0.0
imutils==0.5.4
iopath==0.1.10
ipykernel==6.9.1
ipython==8.3.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.4
lazypredict==0.2.12
lightgbm==3.3.5
markdown==3.4.1
markupsafe==2.1.1
matplotlib-inline==0.1.2
matplotlib==3.6.2
mmcv-full==1.5.0
mmdet==2.26.0
mmpose==0.29.0
moviepy==1.0.3
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
nest-asyncio==1.5.5
networkx==2.8.8
nodeenv==1.7.0
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.1
parso==0.8.3
pathspec==0.10.2
pathtools==0.1.2
pickleshare==0.7.5
pillow==9.2.0
pip==22.3.1
platformdirs==2.5.2
portalocker==2.6.0
pre-commit==2.21.0
proglog==0.1.10
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycocotools==2.0
pycparser==2.21
pygments==2.11.2
pymemcache==4.0.0
pyopenssl==22.1.0
pyparsing==3.0.9
pyreadline3==3.4.1
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywavelets==1.4.1
pywin32==305
pyyaml==6.0
pyzmq==22.3.0
regex==2022.10.31
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==65.6.3
shapely==2.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
sympy==1.11.1
tabulate==0.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchsampler==0.1.2
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing-extensions==4.4.0
urllib3==1.26.11
virtualenv==20.17.1
wandb==0.13.11
wcwidth==0.2.5
werkzeug==2.2.2
wheel==0.38.4
win-inet-pton==1.1.0
xgboost==1.7.4
xtcocotools==1.12
yacs==0.1.8
yapf==0.32.0
zipp==3.10.0